More timestretching improvements
