+++ modules/yup_dsp/yup_dsp.cpp
@@ -86,6 +86,9 @@
 #include "metering/yup_KMeterState.cpp"
 #include "designers/yup_FilterDesigner.cpp"
 #include "convolution/yup_PartitionedConvolver.cpp"
+#include "stretching/yup_TimeStretchProcessorEngine.h"
+#include "stretching/yup_TimeStretchTimeDomainBackend.h"
+#include "stretching/yup_TimeStretchBungeeBackend.h"
 #include "stretching/yup_TimeStretchProcessor.cpp"
 #include "utilities/yup_DspMath.cpp"
